refactor: consolidate validate tests into per-type test files
dedupe attached_object_test.go and attached_object_validate_test.go
by folding hook tests into the former; drop the redundant
TestAOR_RelationshipImmutable table-driven test whose patch-shape
subtests duplicate single-purpose tests in the validate file.

rename the remaining *_validate_test.go files to *_test.go so every
type in pkg/api/v0 follows the same one-file-per-type convention.
